+++ app/Http/Controllers/EmployeeController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Employee;
+use Yajra\DataTables\Facades\DataTables;
+
 
 class EmployeeController extends Controller
 {
@@ -16,6 +18,24 @@ class EmployeeController extends Controller
     {
         $employees = Employee::all();
         return view('employee.all_employee', compact('employees'));
+    }
+
+    public function employeeData(){
+        $employees = Employee::all();
+        $data_table_render= DataTables::of($employees)
+            ->addColumn('DT_RowIndex',function ($row){
+                return '<input type="checkbox" id="qst_id_'.$row["id"].'">';
+            })
+            //add edit and delte option
+                ->addColumn('action',function ($row){
+                    $edit_url=url('employee/edit-employee/'.$row['id']);
+                return '<a href="'.$edit_url.'" class="btn btn-info btn-xs"><i class="far fa-edit"></i></a>'."&nbsp&nbsp;".
+                     '<button onClick="deleteEmployee('.$row['id'].')" class="btn btn-danger btn-xs"><i class="far fa-trash-alt"></i></button>';
+            })
+            ->rawColumns(['DT_RowIndex','action'])
+            ->addIndexColumn()
+            ->make(true);
+        return $data_table_render;
     }
 
     public function storeEmployee(Request $request)
@@ -79,7 +99,11 @@ class EmployeeController extends Controller
     public function deleteEmployee($id)
     {
         $employee = Employee::find($id);
-        $employee->delete();
-        return redirect()->back()->with('danger','Employee Deleted Successfully!');
+        if($employee){
+            $employee->delete();
+            return response()->json('success',201);
+        }else{
+            return response()->json('error',422);
+        }
     }
 }
