+++ app/Http/Controllers/LnameController.php
@@ -7,30 +7,21 @@ use App\Ltype;
 use App\Lgroup;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
+use Yajra\DataTables\Facades\DataTables;
+use Illuminate\Support\Facades\Auth;
 
 class LnameController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function index()
+    public function showAddLedgerName()
     {
         $lnames = DB::table('lnames')
             ->join('lgroups', 'lnames.lgroup_id', '=', 'lgroups.id')
             ->join('ltypes', 'ltypes.id', '=', 'lnames.ltype_id')
             ->select('lnames.*', 'lgroups.name as lgroup_name', 'ltypes.name as ltype_name')
             ->get();
-        //dd($lnames);
-        return view('lname.view', compact('lnames'));
+        //$ladgerName=$lnames;
+        return view('lname.view');
     }
-
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
     public function create()
     {
         $lgroups = Lgroup::all();
@@ -66,6 +57,49 @@ class LnameController extends Controller
         return redirect()->back()->with('success','Ledger Group Added Successfully!');
     }
 
+    public function ladgerNameInfoData(){
+        //$ladgernameinfo = Lname::all();
+        $lnames = DB::table('lnames')
+            ->join('lgroups', 'lnames.lgroup_id', '=', 'lgroups.id')
+            ->join('ltypes', 'ltypes.id', '=', 'lnames.ltype_id')
+            ->select('lnames.*', 'lgroups.name as lgroup_name', 'ltypes.name as ltype_name')
+            ->get();
+            foreach($lnames as $lname){
+                $customData[]=[
+                    'id'=>$lname->id,
+                    'name'=>$lname->name,
+                    'ltype_name'=>$lname->ltype_name,
+                    'lgroup_name'=>$lname->lgroup_name,
+                    'ltype_name'=>$lname->ltype_name,
+                    'unit'=>$lname->unit,
+                    'is_debit'=>$lname->is_debit,
+                ];
+            }
+        $data_table_render= DataTables::of($customData)
+            ->addColumn('DT_RowIndex',function ($row){
+                return '<input type="checkbox" id="qst_id_'.$row["id"].'">';
+            })
+            //add edit and delte option
+                ->addColumn('action',function ($row){
+                    $edit_url=url('ledgername/edit/'.$row['id']);
+                return '<a href="'.$edit_url.'" class="btn btn-info btn-xs"><i class="far fa-edit"></i></a>'."&nbsp&nbsp;".
+                     '<button onClick="destroyladgerName('.$row['id'].')" class="btn btn-danger btn-xs"><i class="far fa-trash-alt"></i></button>';
+            })
+            ->editColumn('is_debit',function ($row){
+                $htmlElement="";
+                 if ($row['is_debit']==1){
+                     $htmlElement='<button class="btn btn-success btn-xs">DR</button>';
+                 }else{
+                     $htmlElement='<button class="btn btn-danger btn-xs">CR</button>';
+                 }
+                 return $htmlElement;
+            })
+            ->rawColumns(['DT_RowIndex','action','is_debit'])
+            ->addIndexColumn()
+            ->make(true);
+        return $data_table_render;
+    }
+
     /**
      * Display the specified resource.
      *
@@ -83,7 +117,7 @@ class LnameController extends Controller
      * @param  \App\Lname  $lname
      * @return \Illuminate\Http\Response
      */
-    public function edit($id)
+    public function editladgerName($id)
     {
         $lname = Lname::find($id);
         $lgroups = Lgroup::all();
@@ -112,7 +146,7 @@ class LnameController extends Controller
 
         $lname->update($request->all());
 
-        return redirect()->route('lname.index')->with('success','Ledger Updated Successfully!');
+        return redirect()->route('showAddLadgerName')->with('success','Ledger Updated Successfully!');
     }
 
     /**
@@ -121,10 +155,19 @@ class LnameController extends Controller
      * @param  \App\Lname  $lname
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Request $request)
+    public function destroyladgerName(Request $request)
     {
         $lname = Lname::findOrFail($request->id);
-        $lname->delete();
-        return redirect()->route('lname.index')->with('danger','Ledger Deleted Successfully!');
+        if($lname){
+          $lname->delete();
+          return response()->json('success',201);
+        }else{
+            return response()->json('error',422);
+            }
+    }
+
+    public function totalLedgerName(){
+        $totalLedgerName=Lname::count();
+        return response()->json($totalLedgerName);
     }
 }
